src/main/webapp/app/entities/filiale/delete/filiale-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { FilialeService } from '../service/filiale.service';

import { FilialeDeleteDialogComponent } from './filiale-delete-dialog.component';

describe('Filiale Management Delete Component', () => {
  let comp: FilialeDeleteDialogComponent;
  let fixture: ComponentFixture<FilialeDeleteDialogComponent>;
  let service: FilialeService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, FilialeDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(FilialeDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(FilialeDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(FilialeService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
